fixes receipt endpoint

+++ src/main/java/org/example/booking/dto/BookingRequest.java
@@ -1,4 +1,4 @@
 package org.example.booking.dto;
 
-public record BookingRequest(Long hotel_id, Long rooms) {
+public record BookingRequest(Long hotel_id, int rooms) {
 }
